Move OnOffenceHandler impl for Pallet to impl.rs
